Add Systemd startup file to set USBFS memory
This is an alternative to the recommended approach of setting this in
the kernel commandline arguments in /etc/default/grub.  Setting at
runtime using a systemd persistent task means that we can 1) activate
this on our first pass without a reboot but still in a way that
reasonably tests that it will work after a reboot, and 2) in a way
that's relatively unobtrusive to other system parameters.  This should
also be relatively easy to extend to work with launch ordering if we
eventually launch some parts of ROS with systemd.